feat(page): details
- Replacement method to stateless widgets.
